gpio: omap: Make GPIO operations fall through pinctrl-single
On OMAP platforms, there are two diferent hardware blocks for
I/O multiplexing / pad configuration and GPIO control. So two
different drivers are used: pinctrl-single and gpio-omap.

Since two separate drivers are used there is no coordination
between these and a I/O pad is not configured as a GPIO when
a GPIO pin is requested.

This patch adds pinctrl back-end commands to the GPIO OMAP
driver so the pinmux_ops functions in the pinctrl-single
driver are called for each GPIO operation.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>