clk: samsung: set SPI0 ioclk and slck as ignore unused
There is no one who handles the ioclk and sclk for the SPI0 bus,
in order to be able to use the SPI 0 bus, set them as ignore
unused by using the CLK_IGNORE_UNUSED flag in the gate
declaration.

The Exynos5433 SoC needs to have all the clocks enabled, they get
all disabled even if one of them is disabled.

Change-Id: I57a625f20c40b4d76046da1a8caf5976454dda1b
Signed-off-by: Andi Shyti <andi.shyti@samsung.com>
Signed-off-by: Hoegeun Kwon <hoegeun.kwon@samsung.com>