drm/exynos/hdmi: workaround to check invalid modes with attached drm bridge
Currently, there is no way to check valid modes for bridge from
user and if the attached drm bridge, mhl adapter, does not support
a mode with fixup, then it is not possible to be set on exynos-hdmi.
So this patch adds workaround to check invalid modes with attached
drm bridge during mode_valid from exynos-hdmi.

Change-Id: I4dcd7edaa1261c1c3ed349c2dcddec226efcc6ec
Signed-off-by: Seung-Woo Kim <sw0312.kim@samsung.com>