drm/exynos: signal pending fence on plane disable
The Exynos DRM driver signals the fence of the previous plane, so the
fences are only signaled if a new plane is set. This means that a
fence will never be signaled if a plane is disabled, making the other
driver waiting for the fence to either time out or wait forever.

Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>