[media] vb2: add implicit dma-buf fence support to core
This patch adds implicit dma-buf fence support to videobuf2 core. The
fences contexts are created on reqbufs and a fence attached to each
buffer plane before buffers are handled to the driver for processing.

The buffers are then marked as done so user-pace is able to do a DQBUF
even while the buffer is still being processed by the driver.

Once the driver finish processing and handles back the buffers to vb2,
then fences for planes are signaled so another driver sharing the DMA
buffers can access them if was waiting for these fences.

Similarly, buffers that are enqueued with QBUF, are only enqueued to
the driver once its pending fences are signaled by the other driver
that may still be accessing the buffers.