sched: replace O(1) scheduler with a simple fair-share scheduler
This patch replaces the O(1) scheduler with a fair-share scheduler,
which switches between tasks by using a simple round robin algorithm
that traverses a multi-level linked list.

The first level of the linked list are the groups with users that
have running processes, the second level are the users for a given
group and the third level are the tasks for a given user.

The algorithm doesn't have SMP support and is likely to be buggy,
but at least works. The processes are scheduled and the system is
stable. Performance is very bad though, specially for workloads
that require low latency.

It was developed as an assignment for an Operating Systems course
at the "Universidad Catolica de Asuncion" in 2005. It was found
by code archaeologist (and one of the authors) Roberto Rodriguez
Alcala.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>