configs: New config in sound/soc for v5.4-rc1
Message-id: <20190918194331.8835-1-jcline@redhat.com>
Patchwork-id: 272493
O-Subject: [ARK INTERNAL PATCH] [redhat] configs: New config in sound/soc for v5.4-rc1
Bugzilla:
RH-Acked-by: John Linville <linville@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* sound/soc/codecs/Kconfig (help): SND_SOC_UDA1334 (tristate)
    The UDA1334 is an NXP audio codec, supports the I2S-bus data format
    and has basic features such as de-emphasis (at 44.1 kHz sampling
    rate) and mute.

Cc: John Linville <linville@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>