[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CHARGER_BQ2515X:

 Say Y to enable support for the TI BQ2515X family of battery
 charging integrated circuits. The BQ2515X are highly integrated
 battery charge management ICs that integrate the most common
 functions for wearable devices, namely a charger, an output voltage
 rail, ADC for battery and system monitoring, and push-button
 controller.

 Symbol: CHARGER_BQ2515X [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:613
   Prompt: TI BQ2515X battery charger family
   Depends on: POWER_SUPPLY [=y] && I2C [=y] && (GPIOLIB [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>