Fedora: cleanup PCMCIA configs, move to x86
None of the Fedora architectures other than x86 support PCMCIA cards
and arm/s390 arches had already disabled the support, as had ARK in
general so move all the configs to reside in fedora/x86 as it's
config would actually be the only consumers of the drivers now.

No functional change to non Fedora configs.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>