[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CHARGER_BQ256XX:

 Say Y to enable support for the TI BQ256XX battery chargers. The
 BQ256XX family of devices are highly-integrated, switch-mode battery
 charge management and system power path management devices for single
 cell Li-ion and Li-polymer batteries.

 Symbol: CHARGER_BQ256XX [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:658
   Prompt: TI BQ256XX battery charger driver
   Depends on: POWER_SUPPLY [=y] && I2C [=y] && (GPIOLIB [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

 CONFIG_CHARGER_LTC4162L:

 Say Y to include support for the Analog Devices (Linear Technology)
 LTC4162-L battery charger connected to I2C.

 Symbol: CHARGER_LTC4162L [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:517
   Prompt: LTC4162-L charger
   Depends on: POWER_SUPPLY [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>