Don't create empty pending-common files on pending-fedora commits
There is a bug in gen_config_patches.sh which will assumes all new
configs are in redhat/configs/pending-common.  This is not true
for pending-fedora configs.  As a result the script reads the
pending-feodra config and tries to copy a pending-common config
into the final location.

The pending-common file does not exist thus creating an empty final
config file.

Fix this by filtering out any pending-fedora configs from the list of
configs to process.

Tested manually.

Signed-off-by: Don Zickus <dzickus@redhat.com>