fedora: configs: Fix WM5102 Kconfig
Commit 0ec8e4cdd877 ("Build Fedora x86s kernels with bytcr-wm5102"),
tried to enable support for audio on x86_64 device with a WM5102 codec.

But it enabled the CONFIG_MFD_ARIZONA symbol, which is not
user-selectable, so setting it in the fedora config is not necessary
and is in essence a no-op.

The CONFIG_MFD_ARIZONA gets automatically selected as necessary by the
CONFIG_MFD_ARIZONA_I2C/CONFIG_MFD_ARIZONA_SPI Kconfig options.

The x86_64 devices with a WM5102 codec, all use the SPI bus, so this
commit enables CONFIG_MFD_ARIZONA_SPI=m. This commit also a whole bunch
of other symbols which depend on CONFIG_MFD_ARIZONA and thus now
become available. Those necessary for supporting these x86_64 devices
are set to 'm' the rest is left disabled.

Starting with 5.14-rc1 all of the Arizona code can be built as a module
so enabling this will not grow the bzImage size.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>