[redhat] New configs in lib/Kconfig.kcsan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KCSAN [=n]
 Type  : bool
 Defined at lib/Kconfig.kcsan:23
   Prompt: KCSAN: dynamic data race detector
   Depends on: HAVE_ARCH_KCSAN [=y] && HAVE_KCSAN_COMPILER [=y] && DEBUG_KERNEL [=y] && !KASAN [=n] && !KCSAN_KCOV_BROKEN [=n]
   Location:
     -> Kernel hacking
       -> Generic Kernel Debugging Instruments
 Selects: STACKTRACE [=y]

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Aristeu Rozanski <arozansk@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Eugene Syromiatnikov <esyromia@redhat.com>
Cc: "Herton R. Krzesinski" <herton@redhat.com>
Cc: Nico Pache <npache@redhat.com>
Cc: Oleg Nesterov <onestero@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>