[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VIDEO_OV5693 [=n]
 Type  : tristate
 Defined at drivers/media/i2c/Kconfig:1062
   Prompt: OmniVision OV5693 sensor support
   Depends on: MEDIA_SUPPORT [=m] && I2C [=y] && VIDEO_V4L2 [=m]
   Visible if: MEDIA_SUPPORT [=m] && I2C [=y] && VIDEO_V4L2 [=m] && MEDIA_CAMERA_SUPPORT [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Multimedia support (MEDIA_SUPPORT [=m])
           -> Media ancillary drivers
             -> Camera sensor devices
 Selects: V4L2_FWNODE [=n]

---

Cc: Jarod Wilson <jarod@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>