spec: make linux-firmware weak(er) dependency
Users of cloud/VM instances want to save up to ~300 MB space,
because they rarely need firmware to be installed.

We already allow linux-firmware to be uninstalled, but problem
is that subsequent updates of kernel package pull it back.

Goal is:
1) Have firmware installed by default, and require user action
   to uninstall it.
2) If uninstalled, it should stay uninstalled after kernel update,
   and kernel can be updated freely without it.
3) If kept installed, minimal required linux-firmware version
   still needs to be enforced.

Make firmware weak dependency [1] (available since rpm 4.13),
but still enforce minimal version if it's installed.

[1] https://docs.fedoraproject.org/en-US/packaging-guidelines/WeakDependencies/

Signed-off-by: Jan Stancek <jstancek@redhat.com>