[redhat] Update scripts to put new configs under ark-common
The current automated workflow creates new kernel config items in a
pending-common/generic directory and produces an MR which proposes
moving them to common/generic.

These are config settings for ARK so really we should use a pending-ark
directory and propose moving them to the ark/generic directory.

Update the relevant scripts and documentation to take care of this.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>