redhat/self-test: Fix BUILD verification test
This test fails because it assumes that the top commit will be a
"[redhat]: kernel-version" commit.  This is true for the main tree but
isn't true for development branches.

Change the test to look for the latest "[redhat]: kernel-version" commit
and compare the value of BUILD to the value in that kernel version.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>