[redhat] New configs in fs/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HUGETLB_PAGE_OPTIMIZE_VMEMMAP_DEFAULT_ON [=n]
 Type  : bool
 Defined at fs/Kconfig:261
   Prompt: Default optimizing vmemmap pages of HugeTLB to on
   Depends on: HUGETLB_PAGE_OPTIMIZE_VMEMMAP [=y]
   Location:
     Main menu
       -> File systems
         -> Pseudo filesystems

---

Cc: Brian Foster <bfoster@redhat.com>
Cc: Carlos Maiolino <cmaiolino@redhat.com>
Cc: Eric Sandeen <esandeen@redhat.com>
Cc: "Bill O'Donnell" <bodonnel@redhat.com>
Cc: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>