[redhat] New configs in fs/dlm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DLM_DEPRECATED_API [=n]
 Type  : bool
 Defined at fs/dlm/Kconfig:12
   Prompt: DLM deprecated API
   Depends on: DLM [=m]
   Location:
     Main menu
       -> File systems
         -> Distributed Lock Manager (DLM) (DLM [=m])

---

Cc: Alexander Aring <aahringo@redhat.com>
Cc: David Teigland <teigland@redhat.com>
Cc: Bob Peterson <rpeterso@redhat.com>
Cc: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>