[redhat] New configs in drivers/vfio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VFIO_PCI_ZDEV_KVM [=y]
 Type  : bool
 Defined at drivers/vfio/pci/Kconfig:47
   Prompt: VFIO PCI extensions for s390x KVM passthrough
   Depends on: VFIO [=m] && PCI [=y] && MMU [=y] && S390 [=y] && KVM [=m]
   Location:
     Main menu
       -> Device Drivers
         -> VFIO Non-Privileged userspace driver framework (VFIO [=m])

---

Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Eric Auger <eric.auger@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_VFIO_PCI_ZDEV_KVM moved under ark/generic/s390x/.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>