redhat: use new genlog.sh script to detect changes for dist-release
The current dist-release process has a flaw, where both its subtargets
(dist-release-changed and dist-release-finish) depends on setup-source
target to run (so genspec.sh is run twice). To avoid having to do the
work twice and do and revert changes before dist-release-finish is
called, we now detect changes by creating the changelog using genlog.sh
first to detect if there are really changes, and do the marker check
using an if condition instead of relying on git to see if there are
changes.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>