drm/ssd130x: Add drm_panic support
Add support for the drm_panic infrastructure, which allows to display
a user friendly message on the screen when a Linux kernel panic occurs.

The display controller doesn't scanout the framebuffer, but instead the
pixels are sent to the device using a transport bus. For this reason, a
.panic_flush handler is needed to flush the panic image to the display.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Series-cc: Jocelyn Falempe <jfalempe@redhat.com>
Series-cc: Maxime Ripard <mripard@kernel.org>